REFAC(crypto): Variable crypto head length
The encrypted output of block cipher will always be the same as
the input length. However, authenticated encryption algorithms
produces "tag" to verify the authenticity of the cipher text and
this "tag" need to be attached to the cipher text.
This tag may vary from cipher to cipher, therefore it is logical
to define a custom "head length" for each cipher.